--- a/interface/src/utils/assets.ts
+++ b/interface/src/utils/assets.ts
@@ -18,3 +18,7 @@ export const makeNetworkAsset = (network: NetworkInfo): BlockchainToken => {
 export const getBalanceRegistryKey = (asset: BlockchainToken) => {
   return `${asset.coin}-${asset.chainId}-${asset.contractAddress.toLowerCase()}`
 }
+
+export const getBalanceRegistryKeyRaw = (coin: number, chainId: string, contract: string) => {
+  return `${coin}-${chainId}-${contract.toLowerCase()}`
+}