|Issue|https://github.com/brave/brave-browser/issues/28634|
-|-

⚠️ This PR is compatible only with brave-core 1.50.x. The MAJOR version has been bumped due to change in interface.

## Comparison of time taken to scan all token balances

### On page load

|   |Before|After|
|--|-------|----|
|**Ethereum**|45s|4s|
|**Polygon**|13s|4s|
|**Optimism**|9s|4s|
|**BSC**|12s|2s|

### On switch network

|   |Before|After|
|--|-------|----|
|**Ethereum**|20s|2s|
|**Polygon**|6s|2s|
|**Optimism**|4s|2s|
|**BSC**|4s|1s|